style(build): rename gulp script file
