+++ scripts/gulp/gulpfile.ts
@@ -5,6 +5,6 @@ import './tasks/bundle/bundle';
 import './tasks/docs/docs';
 import './tasks/copy-sources';
 import './tasks/bump-versions';
-import './tasks/playground/playground';
+import './tasks/playground/playground-schematic';
 
 task('default', ['copy-sources']);
